+++ forui/example/lib/sandbox.dart
@@ -19,11 +19,17 @@ class _SandboxState extends State<Sandbox> {
   Widget build(BuildContext context) => Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-          const FSwitch(
-            label: Text('Airplane Mode'),
-            description: Text('Disable all wireless connections.'),
-            semanticLabel: 'Airplane Mode',
+          FSelectGroup(
+            label: const Text('Select Group'),
+            description: const Text('Select Group Description'),
+            controller: FMultiSelectGroupController(min: 1, max: 2, values: {1}),
+            items: [
+              FSelectGroupItem.checkbox(value: 1, label: const Text('Checkbox 1'), semanticLabel: 'Checkbox 1'),
+              FSelectGroupItem.checkbox(value: 2, label: const Text('Checkbox 2'), semanticLabel: 'Checkbox 2'),
+              FSelectGroupItem.checkbox(value: 3, label: const Text('Checkbox 3'), semanticLabel: 'Checkbox 3'),
+            ],
           ),
+          const SizedBox(height: 20),
           FTooltip(
             longPressExitDuration: const Duration(seconds: 5000),
             tipBuilder: (context, style, _) => const Text('Add to library'),