+++ forui/lib/widgets/select_group.dart
@@ -1,0 +1,10 @@
+/// {@category Widgets}
+///
+/// A group of form fields that represent a single selection.
+///
+/// See https://forui.dev/docs/select-group for working examples.
+library forui.widgets.select_group;
+
+export '../src/widgets/select_group/select_group.dart';
+export '../src/widgets/select_group/select_group_controller.dart';
+export '../src/widgets/select_group/select_group_item.dart';