+++ samples/lib/widgets/checkbox.dart
@@ -7,13 +7,15 @@ import 'package:forui_samples/sample_scaffold.dart';
 
 @RoutePage()
 class CheckboxPage extends SampleScaffold {
+  final bool initialValue;
   final bool enabled;
-  final String? forceErrorText;
+  final String? error;
 
   CheckboxPage({
     @queryParam super.theme,
+    @queryParam this.initialValue = false,
     @queryParam this.enabled = true,
-    @queryParam this.forceErrorText,
+    @queryParam this.error,
   });
 
   @override
@@ -22,16 +24,49 @@ class CheckboxPage extends SampleScaffold {
         children: [
           ConstrainedBox(
             constraints: const BoxConstraints(maxWidth: 290),
-            child: FCheckbox(
-              label: const Text('Accept terms and conditions'),
-              description: const Text('You agree to our terms and conditions.'),
-              semanticLabel: 'Accept terms and conditions',
-              onChange: (value) {}, // Do something.
-              forceErrorText: forceErrorText,
+            child: _Checkbox(
+              initialValue: initialValue,
               enabled: enabled,
+              error: error,
             ),
           ),
         ],
+      );
+}
+
+class _Checkbox extends StatefulWidget {
+  final bool initialValue;
+  final bool enabled;
+  final String? error;
+
+  const _Checkbox({
+    required this.initialValue,
+    required this.enabled,
+    this.error,
+  });
+
+  @override
+  State<_Checkbox> createState() => _CheckboxState();
+}
+
+class _CheckboxState extends State<_Checkbox> {
+  bool state = false;
+
+  @override
+  void initState() {
+    super.initState();
+    state = widget.initialValue;
+  }
+
+  @override
+  Widget build(BuildContext context) => FCheckbox(
+        label: const Text('Accept terms and conditions'),
+        description: const Text('You agree to our terms and conditions.'),
+        error: widget.error != null ? Text(widget.error!) : null,
+        semanticLabel: 'Accept terms and conditions',
+        value: state,
+        onChange: (value) => setState(() => state = value),
+        enabled: widget.enabled,
       );
 }
 
@@ -50,12 +85,29 @@ class RawCheckboxPage extends SampleScaffold {
         children: [
           ConstrainedBox(
             constraints: const BoxConstraints(maxWidth: 300),
-            child: FCheckbox(
-              onChange: (value) {}, // Do something.
-              enabled: enabled,
-            ),
+            child: _RawCheckbox(enabled: enabled),
           ),
         ],
+      );
+}
+
+class _RawCheckbox extends StatefulWidget {
+  final bool enabled;
+
+  const _RawCheckbox({required this.enabled});
+
+  @override
+  State<_RawCheckbox> createState() => _RawCheckboxState();
+}
+
+class _RawCheckboxState extends State<_RawCheckbox> {
+  bool state = false;
+
+  @override
+  Widget build(BuildContext context) => FCheckbox(
+        value: state,
+        onChange: (value) => setState(() => state = value),
+        enabled: widget.enabled,
       );
 }
 
@@ -103,10 +155,19 @@ class _RegisterFormState extends State<RegisterForm> {
               validator: (value) => 8 <= (value?.length ?? 0) ? null : 'Password must be at least 8 characters long.',
             ),
             const SizedBox(height: 15),
-            FCheckbox(
-              label: const Text('Accept terms and conditions'),
-              description: const Text('You agree to our terms and conditions.'),
+            FormField(
+              initialValue: false,
+              onSaved: (value) {
+                // Save values somewhere.
+              },
               validator: (value) => (value ?? false) ? null : 'Please accept the terms and conditions.',
+              builder: (state) => FCheckbox(
+                label: const Text('Accept terms and conditions'),
+                description: const Text('You agree to our terms and conditions.'),
+                error: state.errorText != null ? Text(state.errorText!) : null,
+                value: state.value ?? false,
+                onChange: (value) => state.didChange(value),
+              ),
             ),
             const SizedBox(height: 20),
             FButton(
@@ -116,6 +177,8 @@ class _RegisterFormState extends State<RegisterForm> {
                   // Handle errors here.
                   return;
                 }
+
+                _formKey.currentState!.save();
                 // Do something.
               },
             ),